jwt_proxy/requirements.txt
fastapi
uvicorn
httpx
pyjwt
black
